move exception declerations to the main module
